+++ backend/models/message.go
@@ -1,0 +1,13 @@
+package models
+
+import "github.com/google/uuid"
+
+type Message struct {
+	Base
+	ProjectID uuid.UUID
+	Project   Project
+	UserID    uuid.UUID
+	User      User
+	Role      string
+	Content   string
+}