+++ frontend/.next/static/chunks/app/project/[id]/page-a2a3a9d36fd5108b.js
@@ -1,1 +1,0 @@
-(self.webpackChunk_N_E=self.webpackChunk_N_E||[]).push([[78,105,974],{811:(e,s,_)=>{Promise.resolve().then(_.t.bind(_,3423,23))}},e=>{e.O(0,[423,738,30,358],()=>e(e.s=811)),_N_E=e.O()}]);