+++ frontend/lib/auth.ts
@@ -1,5 +1,12 @@
 import GoogleProvider from "next-auth/providers/google";
-import type { NextAuthOptions } from "next-auth";
+import { NextAuthOptions } from "next-auth";
+
+export type Session = {
+  id: string;
+  name: string;
+  email: string;
+  image: string | null;
+};
 
 export const authOptions: NextAuthOptions = {
   providers: [
@@ -8,10 +15,25 @@ export const authOptions: NextAuthOptions = {
       clientSecret: process.env.GOOGLE_CLIENT_SECRET!,
     }),
   ],
-  secret: process.env.NEXTAUTH_SECRET,
+  pages: {
+    signIn: "/auth/signin",
+    signOut: "/auth/signout",
+    error: "/auth/error",
+    newUser: "/auth/new-user",
+  },
+  secret: process.env.NEXTAUTH_SECRET || "",
+  session: {
+    strategy: "jwt",
+  },
   callbacks: {
     async session({ session, token }) {
-      return session;
+      return {
+        ...session,
+        user: {
+          ...session.user,
+          id: token.sub,
+        },
+      };
     },
   },
 };